chore: refactor of file path
