+++ src/app/(auth)/(no-app-bar))/user/questions/page.tsx
@@ -1,0 +1,3 @@
+export default function Questions() {
+  return <h1>Preguntas</h1>;
+}
