+++ src/app/(auth)/(no-app-bar))/user/saved-items/page.tsx
@@ -1,0 +1,3 @@
+export default function SavedItems() {
+  return <h1>Guardados</h1>;
+}
