+++ src/app/(auth)/(no-app-bar))/user/testimonials/page.tsx
@@ -1,0 +1,3 @@
+export default function Testimonials() {
+  return <h1>Testimonios</h1>;
+}
