--- src/features/user/views/Guardados.tsx
@@ -1,3 +1,0 @@
-export default function Guardados() {
-  return <h1>Guardados</h1>;
-}
