--- src/features/user/views/Preguntas.tsx
@@ -1,3 +1,0 @@
-export default function Preguntas() {
-  return <h1>Preguntas</h1>;
-}
