--- src/features/user/views/Testimonios.tsx
@@ -1,3 +1,0 @@
-export default function Testimonios() {
-  return <h1>Testimonios</h1>;
-}
